+++ plugins/nf-quilt/src/main/nextflow/quilt/QuiltProduct.groovy
@@ -19,14 +19,21 @@ import nextflow.quilt.jep.QuiltPackage
 import nextflow.quilt.nio.QuiltPath
 import nextflow.Session
 
+import java.nio.file.attribute.BasicFileAttributes
 import java.nio.file.Files
+import java.nio.file.FileSystem
+import java.nio.file.FileSystems
+import java.nio.file.FileVisitResult
 import java.nio.file.Path
 import java.nio.file.Paths
+import java.nio.file.PathMatcher
+import java.nio.file.SimpleFileVisitor
 import java.time.LocalDateTime
 
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import groovy.text.GStringTemplateEngine
+import groovy.json.JsonOutput
 
 /**
  * Plugin observer of workflow events
@@ -37,24 +44,26 @@ import groovy.text.GStringTemplateEngine
 @CompileStatic
 class QuiltProduct {
 
-    private final static String KEY_README = 'readme'
     private final static String KEY_META = 'metadata'
+    private final static String KEY_README = 'readme'
     private final static String KEY_SKIP = 'SKIP'
+    private final static String KEY_SUMMARIZE = 'summarize'
 
     /* groovylint-disable-next-line GStringExpressionWithinString */
-    private final static String README_TEMPLATE = '''
+    private final static String DEFAULT_README = '''
 # ${now}
 ## ${msg}
 
 ## workflow
-### scriptFile: ${meta['workflow']['scriptFile']}
-### sessionId: ${meta['workflow']['sessionId']}
+### scriptFile: ${meta['workflow']?.get('scriptFile')}
+### sessionId: ${meta['workflow']?.get('sessionId')}
 - start: ${meta['time_start']}
 - complete: ${meta['time_complete']}
 
 ## processes
 ${meta['workflow']['stats']['processes']}
 '''
+    private final static String DEFAULT_SUMMARIZE = '**.md,**.html'
 
     private final static String[] BIG_KEYS = [
         'nextflow', 'commandLine', 'scriptFile', 'projectDir',
@@ -70,7 +79,7 @@ ${meta['workflow']['stats']['processes']}
 
     static void writeString(String text, QuiltPackage pkg, String filename) {
         String dir = pkg.packageDest()
-        Path path  = Paths.get(dir, filename)
+        Path path  = Paths.get(dir, filename.split('/') as String[])
         try {
             Files.write(path, text.bytes)
         }
@@ -99,15 +108,24 @@ ${meta['workflow']['stats']['processes']}
         this.session = session
         this.pubStatus = -1
         if (session.isSuccess() || pkg.is_force()) {
-            this.pubStatus = publish()
+            try {
+                this.pubStatus = publish()
+            }
+            catch (Exception e) {
+                log.error("publish failed: ${e.getMessage()}", pkg.meta)
+            }
         } else {
             log.info("not publishing: ${pkg} [unsuccessful session]")
         }
     }
 
     int publish() {
-        setupReadme()
-        Map meta = setupMeta()
+        log.debug("publish($msg)")
+        meta = setupMeta()
+        String text = setupReadme()
+        log.debug("setupReadme: $text")
+        Map quilt_summarize = setupSummarize()
+        log.debug("setupSummarize: $quilt_summarize")
         int rc = pkg.push(msg, meta)
         log.info("$rc: pushed package[$pkg] $msg")
         if (rc > 0) {
@@ -120,35 +138,7 @@ ${meta['workflow']['stats']['processes']}
     }
 
     boolean shouldSkip(key) {
-        print("shouldSkip:${key} hasKey:${pkg.meta.containsKey(key)} in ${pkg.meta}\n")
         return pkg.meta.containsKey(key) && pkg.meta[key] == KEY_SKIP
-    }
-
-    String setupReadme() {
-        String text = 'Stub README'
-        try {
-            text = readme()
-        }
-        catch (Exception e) {
-            log.error("setupReadme failed: ${e.getMessage()}", pkg.meta)
-        }
-        if (text != null && text.length() > 0) {
-            log.debug("setupReadme: ${text.length()} bytes")
-            writeString(text, pkg, 'README.md')
-        }
-        return text
-    }
-
-    String readme() {
-        if (shouldSkip(KEY_README)) {
-            log.info("readme=SKIP for ${pkg}")
-            return null
-        }
-        GStringTemplateEngine engine = new GStringTemplateEngine()
-        String raw_readme = pkg.meta_overrides(KEY_README, README_TEMPLATE)
-        log.debug("readme: ${raw_readme}")
-        Writable template = engine.createTemplate(raw_readme).make([meta: meta, msg: msg, now: now()])
-        return template.toString()
     }
 
     Map setupMeta() {
@@ -161,24 +151,32 @@ ${meta['workflow']['stats']['processes']}
         catch (Exception e) {
             log.error("setupMeta failed: ${e.getMessage()}", pkg.meta)
         }
-        writeString("$meta", pkg, 'quilt_metadata.txt')
-        writeString(QuiltPackage.toJson(meta), pkg, 'quilt_metadata.json')
+        writeNextflowMetadata(meta, 'metadata')
         return shouldSkip(KEY_META) ? [:] : meta
     }
 
+    String writeNextflowMetadata(Map map, String suffix) {
+        String filename = "nf-quilt/${suffix}.json"
+        log.debug("writeNextflowMetadata[$suffix]: ${filename}")
+        writeString(QuiltPackage.toJson(map), pkg, filename)
+        return filename
+    }
+
     Map getMetadata(Map cf) {
-        // TODO: Write out config files
         if (cf != null) {
+            cf.remove('executor')
             cf.remove('params')
             cf.remove('session')
-            cf.remove('executor')
+            writeNextflowMetadata(cf, 'config')
+            cf.remove('process')
             printMap(cf, 'config')
         }
         Map params = session.getParams()
         if (params != null) {
+            writeNextflowMetadata(params, 'params')
             params.remove('genomes')
             params.remove('test_data')
-            printMap(params, 'params')
+            //printMap(params, 'params')
         }
         Map wf = session.getWorkflowMetadata().toMap()
         String start = wf['start']
@@ -186,12 +184,13 @@ ${meta['workflow']['stats']['processes']}
         String cmd = wf['commandLine']
         if (wf != null) {
             BIG_KEYS.each { k -> wf[k] = "${wf[k]}" }
+            writeNextflowMetadata(wf, 'workflow')
             wf.remove('container')
             wf.remove('start')
             wf.remove('complete')
             wf.remove('workflowStats')
             wf.remove('commandLine')
-            printMap(wf, 'workflow')
+            //printMap(wf, 'workflow')
             log.info("\npublishing: ${wf['runName']}")
         }
         return [
@@ -202,6 +201,82 @@ ${meta['workflow']['stats']['processes']}
             time_complete: complete,
             workflow: wf,
         ]
+    }
+
+    String setupReadme() {
+        String text = 'Stub README'
+        try {
+            text = readme()
+        }
+        catch (Exception e) {
+            log.error("setupReadme failed: ${e.getMessage()}", pkg.meta)
+        }
+        if (text != null && text.length() > 0) {
+            //log.debug("setupReadme: ${text.length()} bytes")
+            writeString(text, pkg, 'README.md')
+        }
+        return text
+    }
+
+    String readme() {
+        if (shouldSkip(KEY_README)) {
+            log.info("readme=SKIP for ${pkg}")
+            return null
+        }
+        GStringTemplateEngine engine = new GStringTemplateEngine()
+        String raw_readme = pkg.meta_overrides(KEY_README, DEFAULT_README)
+        //log.debug("readme: ${raw_readme}")
+        Writable template = engine.createTemplate(raw_readme).make([meta: meta, msg: msg, now: now()])
+        return template.toString()
+    }
+
+    List<Path> match(String glob) throws IOException {
+        String dir = pkg.packageDest()
+        Path folder = Paths.get(dir)
+        FileSystem fs = FileSystems.getDefault()
+        PathMatcher pathMatcher = fs.getPathMatcher("glob:${glob}")
+        List<Path> matches = []
+
+        Files.walkFileTree(folder, new SimpleFileVisitor<Path>() {
+
+            @Override
+            FileVisitResult visitFile(Path path,
+                    BasicFileAttributes attrs) throws IOException {
+                Path rel = folder.relativize(path)
+                if (pathMatcher.matches(rel)) {
+                    matches.add(rel)
+                }
+                return FileVisitResult.CONTINUE
+            }
+
+            @Override
+            FileVisitResult visitFileFailed(Path file, IOException exc)
+                    throws IOException {
+                return FileVisitResult.CONTINUE
+            }
+
+        })
+        return matches
+    }
+
+    Map setupSummarize() {
+        Map quilt_summarize = [:]
+        if (shouldSkip(KEY_SUMMARIZE)) {
+            return quilt_summarize
+        }
+        String summarize = pkg.meta_overrides(KEY_SUMMARIZE, DEFAULT_SUMMARIZE)
+        String[] wildcards = summarize.split(',')
+        wildcards.each { wildcard ->
+            List<Path> paths = match(wildcard)
+            paths.each { path ->
+                String filename = path.getFileName()
+                quilt_summarize[filename] = path
+            }
+        }
+
+        String qs_json = JsonOutput.toJson(quilt_summarize.keySet() as String[])
+        writeString(qs_json, pkg, 'quilt_summarize.json')
+        return quilt_summarize
     }
 
 }