+++ plugins/nf-quilt/src/main/nextflow/quilt/jep/QuiltParser.groovy
@@ -63,7 +63,7 @@ class QuiltParser {
     }
 
     static QuiltParser forURI(URI uri) {
-        log.debug("forURI[${uri.scheme}] for ${uri}")
+        //log.debug("forURI[${uri.scheme}] for ${uri}")
         if (uri.scheme != SCHEME) {
             String msg =  "Scheme[${uri}] URI:${uri.scheme}] != SCHEME:${SCHEME}"
             throw new IllegalArgumentException(msg)